Address markdown typography review cleanup
